+++ utils/tasks.py
@@ -4,11 +4,16 @@ import subprocess
 def start_celery(worker=True,beat=False):
     from core import core
 
-    cmd = 'pkill celery'
-    subprocess.call(shlex.split(cmd))
+    config = core.api.conf
+    worker  = config.get("celery/worker")
+    beat    = config.get("celery/beat")
+
+    if beat or worker:
+        cmd = 'pkill celery'
+        subprocess.call(shlex.split(cmd))
 
     if worker:
-        log_file = core.config.get('celery/log')
+        log_file = config.get('celery/log')
         cmd = "celery -A tasks.configuration worker --loglevel=info --logfile="+log_file
         print("   > starting celery worker %s"%cmd)
         subprocess.Popen(cmd.split(), shell=False)
